requirements.txt
pyyaml>=6.0.1
jsonschema>=4.19.0
pytest>=7.4.0
pre-commit>=3.3.3
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
crewai>=0.11.0
python-dotenv>=1.0.0
